requirements.txt
markdown2
Jinja2
pyyaml
rich